refactor end-to-end test files
