--- tests/navigation.spec.ts
@@ -1,505 +1,0 @@
-import { test, expect } from '@playwright/test'
-
-test.describe('category navigation', () => {
-	test('user can navigate to a category', async ({ page }) => {
-		await page.goto('/')
-
-		const nav = page.getByRole('navigation')
-		await expect(nav).toBeVisible()
-
-		await nav
-			.getByRole('link', {
-				name: 'Categories',
-				exact: true,
-			})
-			.click()
-
-		await expect(
-			page.getByRole('heading', {
-				name: 'List of categories',
-				exact: true,
-			}),
-		).toBeVisible()
-
-		await expect(page).toHaveURL('/category-list')
-
-		await page
-			.getByRole('link', {
-				name: 'category of commutative rings',
-				exact: true,
-			})
-			.click()
-
-		await expect(
-			page.getByRole('heading', {
-				name: 'category of commutative rings',
-				exact: true,
-			}),
-		).toBeVisible()
-
-		await expect(page).toHaveURL('/category/CRing')
-	})
-
-	test('user can view category details', async ({ page }) => {
-		await page.goto('/category/CRing')
-
-		await expect(
-			page.getByRole('heading', {
-				name: 'category of commutative rings',
-				exact: true,
-			}),
-		).toBeVisible()
-
-		await expect(page.getByText('objects: commutative rings')).toBeVisible()
-		await expect(page.getByText('morphisms: ring homomorphisms')).toBeVisible()
-		await expect(page.getByText('is cocomplete')).toBeVisible()
-		await expect(page.getByText('is locally finitely presentable')).toBeVisible()
-		await expect(page.getByText('is not additive')).toBeVisible()
-		await expect(page.getByText('is not balanced')).toBeVisible()
-		await expect(page.getByText('terminal object: zero ring')).toBeVisible()
-		await expect(page.getByText('coproducts: tensor products')).toBeVisible()
-		await expect(
-			page.getByText('regular epimorphisms: surjective homomorphisms'),
-		).toBeVisible()
-	})
-
-	test('user can open and close a proof for a category', async ({ page }) => {
-		await page.goto('/category/Grp', { waitUntil: 'networkidle' })
-
-		const claim = page.locator('li', { has: page.getByText('is mono-regular') })
-
-		await expect(claim).toBeVisible()
-
-		await claim.locator('button').click()
-
-		await expect(
-			page.getByRole('heading', {
-				name: 'Proof',
-				exact: true,
-			}),
-		).toBeVisible()
-
-		const popup = page.locator('.popup').filter({ hasText: 'Proof' })
-
-		const text = (await popup.textContent())?.trim() ?? ''
-
-		const proof_text = text.replace(/^Proof/, '').trim()
-
-		expect(proof_text.length).toBeGreaterThan(0)
-
-		await popup.getByRole('button', { name: 'close' }).click()
-
-		await expect(
-			page.getByRole('heading', {
-				name: 'Proof',
-				exact: true,
-			}),
-		).not.toBeVisible()
-	})
-
-	test('user can open a derived proof for a category', async ({ page }) => {
-		await page.goto('/category/Set', { waitUntil: 'networkidle' })
-
-		const claim = page.locator('li', { has: page.getByText('has a generator') })
-
-		await expect(claim).toBeVisible()
-
-		await claim.locator('button').click()
-
-		const popup = page.locator('.popup').filter({ hasText: 'Proof' })
-
-		await expect(
-			popup.getByText('Since it is finitary algebraic, it has a generator'),
-		).toBeVisible()
-	})
-
-	test('user can navigate to a category property', async ({ page }) => {
-		await page.goto('/')
-
-		const nav = page.getByRole('navigation')
-		await expect(nav).toBeVisible()
-
-		await nav
-			.getByRole('link', {
-				name: 'Properties',
-				exact: true,
-			})
-			.click()
-
-		await expect(
-			page.getByRole('heading', {
-				name: 'Properties of categories',
-				exact: true,
-			}),
-		).toBeVisible()
-
-		await page
-			.getByRole('link', {
-				name: 'finitely accessible',
-				exact: true,
-			})
-			.click()
-
-		await expect(
-			page.getByRole('heading', {
-				name: 'finitely accessible',
-				exact: true,
-			}),
-		).toBeVisible()
-
-		await expect(page).toHaveURL('/category-property/finitely_accessible')
-	})
-
-	test('user can view category property details', async ({ page }) => {
-		await page.goto('/category-property/finitely_accessible')
-
-		await expect(
-			page.getByRole('heading', {
-				name: 'finitely accessible',
-				exact: true,
-			}),
-		).toBeVisible()
-
-		await expect(page.getByText('A category is finitely accessible if')).toBeVisible()
-
-		await expect(
-			page.getByRole('link', {
-				name: 'accessible',
-				exact: true,
-			}),
-		).toBeVisible()
-
-		const examples = page
-			.getByRole('heading', {
-				name: 'Examples',
-				exact: true,
-			})
-			.locator('xpath=following-sibling::ul[1]')
-
-		await expect(
-			examples.getByRole('link', {
-				name: 'category of fields',
-				exact: true,
-			}),
-		).toBeVisible()
-
-		await expect(
-			examples.getByRole('link', {
-				name: 'category of vector spaces',
-				exact: true,
-			}),
-		).toBeVisible()
-
-		const counterexamples = page
-			.getByRole('heading', {
-				name: 'Counterexamples',
-				exact: true,
-			})
-			.locator('xpath=following-sibling::ul[1]')
-
-		await expect(
-			counterexamples.getByRole('link', {
-				name: 'category of finite sets',
-				exact: true,
-			}),
-		).toBeVisible()
-
-		await expect(
-			counterexamples.getByRole('link', {
-				name: 'category of topological spaces',
-				exact: true,
-			}),
-		).toBeVisible()
-	})
-
-	test('user can navigate to a category implication', async ({ page }) => {
-		await page.goto('/')
-
-		const nav = page.getByRole('navigation')
-		await expect(nav).toBeVisible()
-
-		await nav
-			.getByRole('link', {
-				name: 'Implications',
-				exact: true,
-			})
-			.click()
-
-		await expect(
-			page.getByRole('heading', {
-				name: 'Category implications',
-				exact: true,
-			}),
-		).toBeVisible()
-
-		await page
-			.getByRole('link', {
-				name: 'details',
-			})
-			.first()
-			.click()
-
-		await expect(
-			page.getByRole('heading', {
-				name: 'Implication Details',
-				exact: true,
-			}),
-		).toBeVisible()
-	})
-})
-
-test.describe('structure selector', () => {
-	test('categories are selected by default', async ({ page }) => {
-		await page.goto('/')
-
-		const selector = page
-			.getByRole('combobox', {
-				name: 'Structure',
-				exact: true,
-			})
-			.first()
-
-		await expect(selector).toBeVisible()
-		await expect(selector).toHaveValue('category')
-	})
-
-	test('user can switch to functors', async ({ page }) => {
-		await page.goto('/', { waitUntil: 'networkidle' }) // fix race condition
-
-		const selector = page
-			.getByRole('combobox', {
-				name: 'Structure',
-				exact: true,
-			})
-			.first()
-
-		await expect(selector).toBeVisible()
-		await selector.selectOption('functor')
-
-		await expect(selector).toHaveValue('functor')
-
-		await expect(page).toHaveURL('/functor-list')
-
-		await expect(
-			page.getByRole('heading', {
-				name: 'List of functors',
-				exact: true,
-			}),
-		).toBeVisible()
-	})
-
-	test('functors are selected on a functor route', async ({ page }) => {
-		await page.goto('/functor-properties')
-
-		const selector = page
-			.getByRole('combobox', {
-				name: 'Structure',
-				exact: true,
-			})
-			.first()
-
-		await expect(selector).toBeVisible()
-		await expect(selector).toHaveValue('functor')
-	})
-})
-
-test.describe('functor navigation', () => {
-	test('user can navigate to a functor', async ({ page }) => {
-		await page.goto('/')
-
-		await page
-			.getByRole('link', {
-				name: 'functors',
-				exact: true,
-			})
-			.first()
-			.click()
-
-		const nav = page.getByRole('navigation')
-		await expect(nav).toBeVisible()
-
-		await nav
-			.getByRole('link', {
-				name: 'Functors',
-				exact: true,
-			})
-			.click()
-
-		await expect(
-			page.getByRole('heading', {
-				name: 'List of functors',
-				exact: true,
-			}),
-		).toBeVisible()
-
-		await expect(page).toHaveURL('/functor-list')
-
-		await page
-			.getByRole('link', {
-				name: 'fundamental group functor',
-				exact: true,
-			})
-			.click()
-
-		await expect(
-			page.getByRole('heading', {
-				name: 'fundamental group functor',
-				exact: true,
-			}),
-		).toBeVisible()
-
-		await expect(page).toHaveURL('/functor/pi_1')
-	})
-
-	test('user can view functor details', async ({ page }) => {
-		await page.goto('/functor/pi_1')
-
-		await expect(
-			page.getByRole('heading', {
-				name: 'fundamental group functor',
-				exact: true,
-			}),
-		).toBeVisible()
-
-		await expect(page.getByText('group of homotopy classes of loops')).toBeVisible()
-
-		await expect(page.getByText('preserves products')).toBeVisible()
-		await expect(page.getByText('is essentially surjective')).toBeVisible()
-
-		await expect(page.getByText('is not faithful')).toBeVisible()
-		await expect(page.getByText('does not preserve binary coproducts')).toBeVisible()
-	})
-
-	test('user can open and close a derived proof for a functor', async ({ page }) => {
-		await page.goto('/functor/free_group', { waitUntil: 'networkidle' })
-
-		const claim = page.locator('li', { has: page.getByText('is cocontinuous') })
-
-		await expect(claim).toBeVisible()
-
-		await claim.locator('button').click()
-
-		await expect(
-			page.getByRole('heading', {
-				name: 'Proof',
-				exact: true,
-			}),
-		).toBeVisible()
-
-		const popup = page.locator('.popup').filter({ hasText: 'Proof' })
-
-		await expect(
-			popup.getByText('Since it is a left adjoint, it is cocontinuous'),
-		).toBeVisible()
-
-		await popup.getByRole('button', { name: 'close' }).click()
-
-		await expect(
-			page.getByRole('heading', {
-				name: 'Proof',
-				exact: true,
-			}),
-		).not.toBeVisible()
-	})
-
-	test('user can navigate to a functor property', async ({ page }) => {
-		await page.goto('/functor-list')
-
-		const nav = page.getByRole('navigation')
-		await expect(nav).toBeVisible()
-
-		await nav
-			.getByRole('link', {
-				name: 'Properties',
-				exact: true,
-			})
-			.click()
-
-		await expect(
-			page.getByRole('heading', {
-				name: 'Properties of functors',
-				exact: true,
-			}),
-		).toBeVisible()
-
-		await page
-			.getByRole('link', {
-				name: 'faithful',
-				exact: true,
-			})
-			.click()
-
-		await expect(
-			page.getByRole('heading', {
-				name: 'faithful',
-				exact: true,
-			}),
-		).toBeVisible()
-
-		await expect(page).toHaveURL('/functor-property/faithful')
-	})
-
-	test('user can view functor property details', async ({ page }) => {
-		await page.goto('/functor-property/faithful')
-
-		await expect(
-			page.getByRole('heading', {
-				name: 'faithful',
-				exact: true,
-			}),
-		).toBeVisible()
-
-		await expect(page.getByText('is faithful when')).toBeVisible()
-
-		await expect(
-			page
-				.getByRole('link', {
-					name: 'fully faithful',
-					exact: true,
-				})
-				.first(),
-		).toBeVisible()
-
-		const examples = page
-			.getByRole('heading', {
-				name: 'Examples',
-				exact: true,
-			})
-			.locator('xpath=following-sibling::ul[1]')
-
-		await expect(
-			examples.getByRole('link', {
-				name: 'forgetful functor for rings',
-				exact: true,
-			}),
-		).toBeVisible()
-
-		await expect(
-			examples.getByRole('link', {
-				name: 'free group functor',
-				exact: true,
-			}),
-		).toBeVisible()
-
-		const counterexamples = page
-			.getByRole('heading', {
-				name: 'Counterexamples',
-				exact: true,
-			})
-			.locator('xpath=following-sibling::ul[1]')
-
-		await expect(
-			counterexamples.getByRole('link', {
-				name: 'abelianization functor for groups',
-				exact: true,
-			}),
-		).toBeVisible()
-
-		await expect(
-			counterexamples.getByRole('link', {
-				name: 'fundamental group functor',
-				exact: true,
-			}),
-		).toBeVisible()
-	})
-})
